Test cases added

+++ src/main/java/com/doctor/entities/HospitalAddress.java
@@ -14,7 +14,7 @@ public class HospitalAddress {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer addressId;
+    private Long addressId;
 
     private String country;
